Feature/arduino renzo

+++ cgi-bin/api/actuator.c
@@ -2,15 +2,19 @@
 #include "../library/domoticaTi.h"
 #include <mariadb/mysql.h>
 
-
 int validateActuator(struct actuator);
 void createInsertQueryActuator(char*, struct actuator);
+void createUpdateQueryActuator(char*, struct actuator);
 struct actuator readActuatorJSON();
+int validateActuatorSettings(struct actuator);
+void createUpdateQueryActuatorSettings(char*, struct actuator);
+void actuatorHistory(char*, unsigned int, int);
 
 int CONTENT_SIZE = 0;
 
 int main(int argc, const char* argv[], char* env[]) {
-
+  char username[100];
+  const int acceslevel = searchLogin(env, username);
 
   char METHOD[10];
   CONTENT_SIZE = getContentSize(env);
@@ -40,26 +44,87 @@ int main(int argc, const char* argv[], char* env[]) {
     }
   } else if(argc == 2) { // one data
     if(strcmp(METHOD, "DELETE") == 0) {
+      if(acceslevel >= ACCESLEVEL_DELETE) {
+        const char* actuatorid = argv[1];
 
-      const char* actuatorid = argv[1];
+        if(atoi(actuatorid) != 0 && strlen(actuatorid) > 0) {
+          char* query = malloc(150);
+          sprintf(query, "DELETE FROM actuator WHERE actuatorid=%d", atoi(actuatorid));
+          executeQuery(query);
 
-      if(atoi(actuatorid) != 0 && strlen(actuatorid) > 0) {
-        char* query = malloc(100);
-        sprintf(query, "DELETE FROM actuator WHERE actuatorid=%d", atoi(actuatorid));
-        executeQuery(query);
-
+        } else {
+          errorResponse(400, "validation vailed");
+        }
       } else {
+        errorResponse(455, "Unauthorized, Acceslevel Not High Enough!");
+      }
+
+    } else if(strcmp(METHOD, "GET") == 0) {
+			const char* actuatorid = argv[1];
+      if(atoi(actuatorid) != 0 && strlen(actuatorid) > 0) {
+				char* query = malloc(150);
+        sprintf(query, "SELECT * FROM actuator WHERE actuatorid = %d", atoi(actuatorid));
+				selectQueryJSON(query);
+			} else {
         errorResponse(400, "validation vailed");
       }
-    } else if(strcmp(METHOD, "PUT") == 0) {
+
+		} else if(strcmp(METHOD, "PUT") == 0) {
       struct actuator actuator = readActuatorJSON();
-      errorResponse(501, "This function will be implementend soon!");
-      // TODO MAKE THE UPDATE
+
+			const char* actuatorid = argv[1];
+      if(atoi(actuatorid) != 0 && strlen(actuatorid) > 0 && validateActuator(actuator) > 0) {
+				actuator.actuatorid = atoi(actuatorid);
+				char* query = malloc(200);
+        createUpdateQueryActuator(query, actuator);
+        executeQuery(query);
+			} else {
+        errorResponse(400, "validation vailed");
+      }
     } else {
       errorResponse(400, "check request url");
     }
   } else if(argc > 2) { // one redirect + 1 data
-    errorResponse(404, "url Not Found. Please check all parameters");
+    if(strcmp(METHOD, "PUT") == 0 && strcmp(argv[1], "settings") == 0) {
+      struct actuator actuator = readActuatorJSON();
+
+			const char* actuatorid = argv[2];
+      if(atoi(actuatorid) != 0 && strlen(actuatorid) > 0 && validateActuator(actuator) > 0 && validateActuatorSettings(actuator) > 0) {
+				actuator.actuatorid = atoi(actuatorid);
+				char* query = malloc(300);
+        createUpdateQueryActuatorSettings(query, actuator);
+        executeQuery(query);
+			} else {
+        errorResponse(400, "validation vailed");
+      }
+    } else if(strcmp(METHOD, "PUT") == 0 && strcmp(argv[1], "value") == 0) {
+      struct actuator actuator = readActuatorJSON();
+
+			const char* actuatorid = argv[2];
+      if(atoi(actuatorid) != 0 && strlen(actuatorid) > 0 && validateActuator(actuator) > 0 && validateActuatorSettings(actuator) > 0) {
+				actuator.actuatorid = atoi(actuatorid);
+				char* query = malloc(300);
+        actuatorHistory(username, actuator.actuatorid, actuator.value);
+        sprintf(query, "UPDATE actuator SET value = %d WHERE actuatorid = %d", actuator.value, actuator.actuatorid);
+        executeQuery(query);
+        char staticIpArduino[100];
+        char queryStaticIpArduino[500];
+        sprintf(queryStaticIpArduino, "SELECT staticip FROM arduino WHERE arduinoid = '%d'", actuator.arduinoid);
+        getOneRecordOneColumn(queryStaticIpArduino, staticIpArduino);
+        pingArduino(staticIpArduino, actuator.arduinocomponentid);
+			} else {
+        errorResponse(400, "validation vailed");
+      }
+    } else if(strcmp(METHOD, "GET") == 0 && strcmp(argv[1], "last") == 0) {
+      const char* actuatorid = argv[2];
+        if(atoi(actuatorid) != 0 && strlen(actuatorid) > 0) {
+          char* query = malloc(150);
+          sprintf(query, "SELECT * FROM actuatorhistory WHERE actuatorid = %d ORDER BY date DESC, time DESC LIMIT 25", atoi(actuatorid));
+          selectQueryJSON(query);
+        }
+    } else {
+      errorResponse(404, "URL Not Found. Please check all parameters");
+    }
   } else {
     errorResponse(404, "URL Not found. Please check all parameters");
   }
@@ -76,25 +141,23 @@ int validateActuator(struct actuator actuator) {
   }
   
   if(actuator.arduinoid > 0) { //uncomment following piece when arduino table is active.
-    /*char* query = malloc(100);
-    char* response = malloc(2000);
-    sprintf(query, "SELECT id FROM arduino WHERE id = %d", actuator.actuatorid);
-    if(selectQueryJSON(response, query) <= 0) {
-      return -1;
-    }*/
+    char* query = malloc(100);
+    sprintf(query, "SELECT count(*) FROM arduino WHERE arduinoid = %d", actuator.arduinoid);
+    if(!countRecords(query)) {
+      strncpy(response[1][1], "false", 50);
+      validation--;
+    }
   } else {
     strncpy(response[1][1], "false", 50);
     validation--;
   }
-
-  actuator.value = 0; //No matter what the user sends, the first value is always 0
 
   if(strlen(actuator.type) == 0) {
     strncpy(response[3][1], "false", 50);
     validation--;
   }
 
-  if(strlen(actuator.arduinovalueid) != 3) {
+  if(strlen(actuator.arduinocomponentid) != 3) {
     strncpy(response[4][1], "false", 50);
     validation--;
   }
@@ -116,33 +179,98 @@ int validateActuator(struct actuator actuator) {
   return 1;
 }
 
+int validateActuatorSettings(struct actuator actuator) {
+  short validation = 1, index = 0;
+  char response[ACTUATOR_FIELDS][2][50];
+  for(index = 0; index < ACTUATOR_FIELDS; index++) {
+    strncpy(response[index][0], ACTUATOR_FIELD_NAMES[index], 50);
+    strncpy(response[index][1], "true", 50);
+  }
+  if(! (actuator.iopin == 1 || actuator.iopin == 0)) {
+    strncpy(response[6][1], "false", 50);
+    validation--;
+  }
+
+  if(actuator.maximumvalue <= actuator.minimumvalue) {
+    strncpy(response[7][1], "false", 50);
+    strncpy(response[8][1], "false", 50);
+    validation--;
+  }
+
+  if(actuator.value > actuator.maximumvalue || actuator.value < actuator.minimumvalue) {
+    strncpy(response[2][1], "false", 50);
+    strncpy(response[7][1], "false", 50);
+    strncpy(response[8][1], "false", 50);
+    validation--;
+  }
+
+  if(validation < 1) {
+    printf("STATUS: 400 \ncontent-type: application/json\n\n{");
+    for (index = 0; index < ACTUATOR_FIELDS; index++) {
+      printf("\"%s\":%s%c", response[index][0], response[index][1], index+1 < ACTUATOR_FIELDS ? ',' : ' ');
+    }
+    printf("}");
+    exit(0);
+  }
+
+  return validation;
+}
+
 void createInsertQueryActuator(char* query, struct actuator actuator) {
-  strcpy(query, "INSERT INTO actuator (arduinoid, value, type, arduinovalueid, actuatorname) VALUES(");
+  strcpy(query, "INSERT INTO actuator (arduinoid, type, arduinocomponentid, actuatorname) VALUES(");
 
   sprintf(query, "%s%d", query, actuator.arduinoid);
-  strcat(query, ",");
-  sprintf(query, "%s%d", query, actuator.value);
   strcat(query, ",'");
   strcat(query, actuator.type);
   strcat(query, "','");
-  strcat(query, actuator.arduinovalueid);
+  strcat(query, actuator.arduinocomponentid);
   strcat(query, "','");
   strcat(query, actuator.actuatorname);
   strcat(query, "')");
 }
 
+void createUpdateQueryActuator(char* query, struct actuator actuator) {
+	strcpy(query, "UPDATE actuator SET arduinoid = ");
+
+  sprintf(query, "%s%d", query, actuator.arduinoid);
+  strcat(query, ", type = '");
+  strcat(query, actuator.type);
+  strcat(query, "', arduinocomponentid = '");
+  strcat(query, actuator.arduinocomponentid);
+  strcat(query, "', actuatorname = '");
+  strcat(query, actuator.actuatorname);
+  strcat(query, "' WHERE actuatorid = ");
+	sprintf(query, "%s%d", query, actuator.actuatorid);
+}
+
+void createUpdateQueryActuatorSettings(char* query, struct actuator actuator) {
+	strcpy(query, "UPDATE actuator SET iopin = ");
+
+  sprintf(query, "%s%d", query, actuator.iopin);
+  strcat(query, ", minimumvalue = ");
+  sprintf(query, "%s%d", query, actuator.minimumvalue);
+  strcat(query, ", maximumvalue = ");
+  sprintf(query, "%s%d", query, actuator.maximumvalue);
+  strcat(query, ", value = ");
+  sprintf(query, "%s%d", query, actuator.value);
+  strcat(query, " WHERE actuatorid = ");
+	sprintf(query, "%s%d", query, actuator.actuatorid);
+}
+
+
 struct actuator readActuatorJSON() {
-  struct actuator newActuator = {-1,-1,-1,"","",""}; //define all elements so we can validate the struct
+  struct actuator newActuator = {-1,-1,0,"","","", 0, 0, 0}; //define all elements so we can validate the struct
   
   char* data = malloc(CONTENT_SIZE+10);
   int index;
 
   fgets(data, CONTENT_SIZE, stdin);
   for(index = 0; index < ACTUATOR_FIELDS; index++) {
-    char* dataPointer = strstr(data, ACTUATOR_FIELD_NAMES[index]);
+    char searchString[50];
+    sprintf(searchString, "\"%s\"", ACTUATOR_FIELD_NAMES[index]);
+    char* dataPointer = strstr(data, searchString);
     if(dataPointer != NULL) {
-
-      dataPointer = dataPointer + 3 + strlen(ACTUATOR_FIELD_NAMES[index]); //+3 for: ": "
+      dataPointer = dataPointer + 4 + strlen(ACTUATOR_FIELD_NAMES[index]); //+3 for: ": "
       char* dataPointerEnd = strchr(dataPointer, '"');
 
       int dataSize = dataPointerEnd - dataPointer;
@@ -151,12 +279,11 @@ struct actuator readActuatorJSON() {
       strncpy(retrievedData, dataPointer, dataSize);
       retrievedData[dataSize] = '\0';
       removeBadCharacters(retrievedData);
-
       switch (index) { // starts with 1 because user can't make there own id.
       case 1:
         newActuator.arduinoid = atoi(retrievedData);
         break;
-      
+
       case 2:
         newActuator.value = atoi(retrievedData);
         break;
@@ -166,11 +293,23 @@ struct actuator readActuatorJSON() {
         break;
       
       case 4:
-        strncpy(newActuator.arduinovalueid, retrievedData, ACTUATOR_FIELD_ARDUINOVALUEID_SIZE);
+        strncpy(newActuator.arduinocomponentid, retrievedData, ACTUATOR_FIELD_arduinocomponentid_SIZE);
         break;
       
       case 5:
         memcpy(newActuator.actuatorname, retrievedData, ACTUATOR_FIELD_ACTUATORNAME_SIZE);
+        break;
+      
+      case 6:
+        newActuator.iopin = atoi(retrievedData);
+        break;
+      
+      case 7:
+        newActuator.minimumvalue = atoi(retrievedData);
+        break;
+      
+      case 8:
+        newActuator.maximumvalue = atoi(retrievedData);
         break;
       
       default:
@@ -180,4 +319,10 @@ struct actuator readActuatorJSON() {
   }
 
   return newActuator;
+}
+
+void actuatorHistory(char* username, unsigned int actuatorid, int value) {
+  char query[2000];
+  sprintf(query, "INSERT INTO actuatorhistory(actuatorid, value, bywho, date, time) VALUES (%d, %d, '<b>User: </b>%s', CURDATE() + 0, CURTIME() + 0)", actuatorid, value, username);
+  executeQueryNoOutput(query);
 }
